pyproject.toml
[tool.poetry]
name = "osnap-client"
version = "0.1.1a2"
description = ""
authors = ["Forrest Murray <FMurray@gmail.com>", "Vlad Samoilov <nicelir1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = ">=0.0.153,<0.1.0"
fastapi = ">=0.95.1,<0.96.0"
uvicorn = ">=0.22.0,<0.23.0"
pydantic = ">=1.10.7,<1.11.0"
openai = ">=0.27.5,<0.28.0"
redis = ">=4.5.4,<4.6.0"
cryptography = ">=40.0.2,<40.1.0"
rsa = ">=4.9,<5.0"
httpx = ">=0.24.0,<0.25.0"
watchfiles = "^0.19.0"
networkx = "^3.1"
websockets = "^11.0.2"
python-dotenv = "^1.0.0"
discord = "^2.2.3"
jsonschema = "^4.17.3"
pytest = "^7.3.1"
ipykernel = "^6.23.0"
tiktoken = "^0.4.0"
aiologger = "^0.7.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
autodoc_pydantic = "^1.8.0"
myst_parser = "^0.18.1"
nbsphinx = "^0.8.9"
sphinx = "^4.5.0"
sphinx-autobuild = "^2021.3.14"
sphinx_book_theme = "^0.3.3"
sphinx_rtd_theme = "^1.0.0"
sphinx-typlog-theme = "^0.8.0"
sphinx-panels = "^0.6.0"
toml = "^0.10.2"
myst-nb = "^0.17.1"
linkchecker = "^10.2.1"
sphinx-copybutton = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
